+++ modules/string/src/include/StringJoin.hpp
@@ -7,20 +7,15 @@
 // SPDX-License-Identifier: LGPL-3.0-or-later
 // LICENCE_BLOCK_END
 //=============================================================================
-#include "Banner.hpp"
-#include "NelsonPrint.hpp"
+#pragma once
+//=============================================================================
+#include "ArrayOf.hpp"
+#include "nlsString_exports.h"
 //=============================================================================
 namespace Nelson {
-void
-Banner()
-{
-    NelsonPrint(L"\n");
-    NelsonPrint(L"    _   __     __\n");
-    NelsonPrint(L"   / | / /__  / /________  ____\n");
-    NelsonPrint(L"  /  |/ / _ \\/ / ___/ __ \\/ __ \\\n");
-    NelsonPrint(L" / /|  /  __/ (__  ) /_/ / / / /\n");
-    NelsonPrint(L"/_/ |_/\\___/_/____/\\____/_/ /_/\n");
-}
 //=============================================================================
-} // namespace Nelson
+NLSSTRING_IMPEXP ArrayOf
+StringJoin(const ArrayOf& A, const ArrayOf& delimiter, size_t dimension);
+//=============================================================================
+}
 //=============================================================================